Separate autogen msgpack implementation into separate file
