+++ main/autogen/autogen.cc
@@ -1,16 +1,10 @@
-// has to go first because it violates our poisons
-#include "mpack/mpack.h"
-
-#include "absl/strings/str_split.h"
+#include "main/autogen/autogen.h"
 #include "ast/Helpers.h"
 #include "ast/ast.h"
 #include "ast/treemap/treemap.h"
-#include "common/FileOps.h"
 #include "common/formatting.h"
-#include "common/typecase.h"
-#include "core/Names.h"
-#include "main/autogen/autogen.h"
 #include "main/autogen/autoloader.h"
+#include "main/autogen/msgpack.h"
 
 #include "CRC.h"
 
@@ -379,279 +373,6 @@ string ParsedFile::toString(const core::GlobalState &gs) const {
     }
     return to_string(out);
 }
-
-class MsgpackWriter {
-private:
-    void packName(core::NameRef nm) {
-        u4 id;
-        auto it = symbolIds.find(nm);
-        if (it == symbolIds.end()) {
-            id = symbols.size();
-            symbols.emplace_back(nm);
-            symbolIds[nm] = id;
-        } else {
-            id = it->second;
-        }
-        mpack_write_u32(&writer, id);
-    }
-
-    void packNames(vector<core::NameRef> &names) {
-        mpack_start_array(&writer, names.size());
-        for (auto nm : names) {
-            packName(nm);
-        }
-        mpack_finish_array(&writer);
-    }
-
-    void packString(string_view str) {
-        mpack_write_str(&writer, str.data(), str.size());
-    }
-
-    void packBool(bool b) {
-        if (b) {
-            mpack_write_true(&writer);
-        } else {
-            mpack_write_false(&writer);
-        }
-    }
-
-    void packReferenceRef(ReferenceRef ref) {
-        if (!ref.exists()) {
-            mpack_write_nil(&writer);
-        } else {
-            mpack_write_u16(&writer, ref.id());
-        }
-    }
-
-    void packDefinitionnRef(DefinitionRef ref) {
-        if (!ref.exists()) {
-            mpack_write_nil(&writer);
-        } else {
-            mpack_write_u16(&writer, ref.id());
-        }
-    }
-
-    void packRange(u4 begin, u4 end) {
-        mpack_write_u64(&writer, ((u8)begin << 32) | end);
-    }
-
-    void packDefinition(core::Context ctx, ParsedFile &pf, Definition &def) {
-        mpack_start_array(&writer, def_attrs[version].size());
-
-        // raw_full_name
-        auto raw_full_name = pf.showFullName(ctx, def.id);
-        packNames(raw_full_name);
-
-        // type
-        mpack_write_u8(&writer, static_cast<u8>(def.type));
-
-        // defines_behavior
-        packBool(def.defines_behavior);
-
-        // is_empty
-        packBool(def.is_empty);
-
-        // parent_ref
-        packReferenceRef(def.parent_ref);
-
-        // aliased_ref
-        packReferenceRef(def.aliased_ref);
-
-        // defining_ref
-        packReferenceRef(def.defining_ref);
-        mpack_finish_array(&writer);
-    }
-
-    void packReference(core::Context ctx, ParsedFile &pf, Reference &ref) {
-        mpack_start_array(&writer, ref_attrs[version].size());
-
-        // scope
-        packDefinitionnRef(ref.scope.id());
-
-        // name
-        packNames(ref.name);
-
-        // nesting
-        mpack_start_array(&writer, ref.nesting.size());
-        for (auto &scope : ref.nesting) {
-            packDefinitionnRef(scope.id());
-        }
-        mpack_finish_array(&writer);
-
-        // expression_range
-        auto expression_range = ref.definitionLoc.position(ctx);
-        packRange(expression_range.first.line, expression_range.second.line);
-        // expression_pos_range
-        packRange(ref.loc.beginPos(), ref.loc.endPos());
-
-        // resolved
-        if (ref.resolved.empty()) {
-            mpack_write_nil(&writer);
-        } else {
-            packNames(ref.resolved);
-        }
-
-        // is_defining_ref
-        packBool(ref.is_defining_ref);
-
-        // parent_of
-        packDefinitionnRef(ref.parent_of);
-        mpack_finish_array(&writer);
-    }
-
-    static int assert_valid_version(int version) {
-        if (version < MIN_VERSION || version > MAX_VERSION) {
-            Exception::raise("msgpack version {} not in available range [{}, {}]", version, MIN_VERSION, MAX_VERSION);
-        }
-        return version;
-    }
-
-public:
-    constexpr static int MIN_VERSION = 2;
-    constexpr static int MAX_VERSION = 2;
-
-    // symbols[0..3] are reserved for the Type aliases
-    MsgpackWriter(int version)
-        : version(assert_valid_version(version)), ref_attrs(ref_attr_map.at(version)),
-          def_attrs(def_attr_map.at(version)), symbols(4) {}
-
-    string pack(core::Context ctx, ParsedFile &pf) {
-        char *data;
-        size_t size;
-        mpack_writer_init_growable(&writer, &data, &size);
-        mpack_start_array(&writer, 6);
-
-        mpack_write_true(&writer); // did_resolution
-        packString(pf.path);
-        mpack_write_u32(&writer, pf.cksum);
-
-        // requires
-        mpack_start_array(&writer, pf.requires.size());
-        for (auto nm : pf.requires) {
-            packString(nm.data(ctx)->show(ctx));
-        }
-        mpack_finish_array(&writer);
-
-        mpack_start_array(&writer, pf.defs.size());
-        for (auto &def : pf.defs) {
-            packDefinition(ctx, pf, def);
-        }
-
-        mpack_finish_array(&writer);
-        mpack_start_array(&writer, pf.refs.size());
-        for (auto &ref : pf.refs) {
-            packReference(ctx, pf, ref);
-        }
-        mpack_finish_array(&writer);
-        mpack_finish_array(&writer);
-
-        mpack_writer_destroy(&writer);
-        auto body = string(data, size);
-        MPACK_FREE(data);
-
-        // write header
-        mpack_writer_init_growable(&writer, &data, &size);
-
-        mpack_start_map(&writer, 5);
-
-        packString("symbols");
-        int i = -1;
-        mpack_start_array(&writer, symbols.size());
-        for (auto sym : symbols) {
-            ++i;
-            string str;
-            switch ((Definition::Type)i) {
-                case Definition::Type::Module:
-                    str = "module";
-                    break;
-                case Definition::Type::Class:
-                    str = "class";
-                    break;
-                case Definition::Type::Casgn:
-                    str = "casgn";
-                    break;
-                case Definition::Type::Alias:
-                    str = "alias";
-                    break;
-                default:
-                    str = sym.data(ctx)->show(ctx);
-            }
-            packString(str);
-        }
-        mpack_finish_array(&writer);
-
-        packString("ref_count");
-        mpack_write_u32(&writer, pf.refs.size());
-        packString("def_count");
-        mpack_write_u32(&writer, pf.defs.size());
-
-        packString("ref_attrs");
-        mpack_start_array(&writer, ref_attrs.size());
-        for (auto attr : ref_attrs) {
-            packString(attr);
-        }
-        mpack_finish_array(&writer);
-
-        packString("def_attrs");
-        mpack_start_array(&writer, def_attrs.size());
-        for (auto attr : def_attrs) {
-            packString(attr);
-        }
-        mpack_finish_array(&writer);
-
-        mpack_write_object_bytes(&writer, body.data(), body.size());
-
-        mpack_writer_destroy(&writer);
-
-        auto ret = string(data, size);
-        MPACK_FREE(data);
-
-        return ret;
-    }
-
-private:
-    int version;
-    const vector<string> &ref_attrs;
-    const vector<string> &def_attrs;
-    mpack_writer_t writer;
-
-    vector<core::NameRef> symbols;
-    UnorderedMap<core::NameRef, u4> symbolIds;
-
-    static const map<int, vector<string>> ref_attr_map;
-    static const map<int, vector<string>> def_attr_map;
-};
-
-const map<int, vector<string>> MsgpackWriter::ref_attr_map{
-    {
-        2,
-        {
-            "scope",
-            "name",
-            "nesting",
-            "expression_range",
-            "expression_pos_range",
-            "resolved",
-            "is_defining_ref",
-            "parent_of",
-        },
-    },
-};
-
-const map<int, vector<string>> MsgpackWriter::def_attr_map{
-    {
-        2,
-        {
-            "raw_full_name",
-            "type",
-            "defines_behavior",
-            "is_empty",
-            "parent_ref",
-            "aliased_ref",
-            "defining_ref",
-        },
-    },
-};
 
 string ParsedFile::toMsgpack(core::Context ctx, int version) {
     MsgpackWriter write(version);
